requirements.txt
openai
langchain
pdfplumber
pandas
streamlit
pydantic
scikit-learn
python-dotenv
chromadb
faiss-cpu
PyYAML
tqdm
fastapi
uvicorn
pymupdf
numpy
plotly
sentence-transformers